+++ code/include/boids2D/RootedBoid.hpp
@@ -31,7 +31,7 @@ class RootedBoid : public Boid
    * @param[in] location Location of the object
    * @param[in] t Type of the RootedBoid
    */
-  RootedBoid(glm::vec3 location, BoidType t, float amountFood);
+  RootedBoid(glm::vec3 location, BoidType t, int amountFood);
 
  private:
   float m_radius; ///< Radius of the boid to have "place" to live
